% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpd.R
\name{FitGpdFlex}
\alias{FitGpdFlex}
\title{Fits an GPD to data}
\usage{
FitGpdFlex(data, xpar, fpar, start, method = "Nelder-Mead", ...)
}
\arguments{
\item{data}{The data which should be modeled}

\item{xpar}{The covariates used in}

\item{fpar}{the model}

\item{start}{Vector of length numberOfParamaters}

\item{method}{The optimization method to be used (In case of a univariate
optimization, the *lower* and *upper* boundaries have to be specified.)}

\item{...}{Further arguments passed to optim}
}
\value{
The fitted parameters
}
\description{
Fits an GPD to data
}
