% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGev.R
\name{FitGevFlex}
\alias{FitGevFlex}
\title{Fit flexible GEV}
\usage{
FitGevFlex(data, start, fpar, xpar, likelihood = "standard", std.err = TRUE,
  ...)
}
\arguments{
\item{data}{the data (in case of likelihood = "range" this should be a
matrix or data.frame with two columns)}

\item{start}{the start values}

\item{fpar}{the parameter model}

\item{xpar}{the covariates used in fpar}

\item{likelihood}{either standard or grouped}

\item{std.err}{should standard errors be returned}

\item{...}{additional parameters passed to the optimization}
}
\description{
Fit flexible GEV
}
