% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpd.R
\name{FitGpdFlex}
\alias{FitGpdFlex}
\title{Fits an GPD to data}
\usage{
FitGpdFlex(data, xpar, fpar, start, likelihood = "standard", ...)
}
\arguments{
\item{data}{the data (in case of likelihood = "range" this should be a
matrix or data.frame with two columns)}

\item{xpar}{the covariates used in fpar}

\item{fpar}{the parameter model}

\item{start}{the start values}

\item{likelihood}{either standard or grouped}

\item{...}{Further arguments passed to optim}
}
\value{
The fitted parameters
}
\description{
Fits an GPD to data
}
